LayoutTests/accessibility/ios-simulator/color-well-expected.txt


This test thats color input types are accessible on iOS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.isIgnored is false
PASS element.stringAttributeValue('AXRoleDescription') is 'color well'
PASS element.stringAttributeValue('AXColorStringDescription') is 'rgb 0.90196 0.39216 0.39608 1'
PASS successfullyParsed is true

TEST COMPLETE
